+++ lib/spreadsheet_architect/class_methods/csv.rb
@@ -3,8 +3,8 @@ require 'csv'
 module SpreadsheetArchitect
   module ClassMethods
     def to_csv(opts={})
-      opts = SpreadsheetArchitect::Utils.get_cell_data(opts, self)
-      options = SpreadsheetArchitect::Utils.get_options(opts, self)
+      opts = SpreadsheetArchitect::Utils.get_options(opts, self)
+      options = SpreadsheetArchitect::Utils.get_cell_data(opts, self)
 
       CSV.generate do |csv|
         if options[:headers]